# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Auto Assessor API
  description: Vehicle damage assessment service powered by AI vision
  contact:
    name: Nedbank Avo
    email: avo@nedbank.co.za
  version: 1.0.0
servers:
- url: http://localhost:8085
  description: Generated server url
...
paths:
  /api/admin/assessments/{assessmentId}:
    get:
      tags:
      - Admin Assessment Management
      summary: Get assessment details
      description: "Retrieves a single assessment summary for admin workflows, including\
        \ status, image counters, token usage, and timestamps."
      operationId: getAssessmentDetails
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Assessment summary returned successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
    put:
      tags:
      - Admin Assessment Management
      summary: Update assessment
      description: Updates editable assessment data such as vehicle details and odometer
        reading.
      operationId: updateAssessment
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssessmentRequest"
        required: true
      responses:
        "200":
          description: Assessment updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
    delete:
      tags:
      - Admin Assessment Management
      summary: Archive assessment
      description: Soft deletes an assessment by marking it as archived. The record
        remains available for restore operations.
      operationId: deleteAssessment
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Assessment archived successfully
        "404":
          description: Assessment not found
...
  /api/admin/assessments/{assessmentId}/restore:
    post:
      tags:
      - Admin Assessment Management
      summary: Restore assessment
      description: Restores a previously archived assessment back into active workflows.
      operationId: restoreAssessment
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Assessment restored successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
  /api/admin/assessments/{assessmentId}/images/{angle}/retry:
    post:
      tags:
      - Admin Assessment Management
      summary: Retry image analysis
      description: Marks an image angle for reprocessing when the previous analysis
        failed.
      operationId: retryImageAnalysis
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      - name: angle
        in: path
        description: Vehicle image angle to retry
        required: true
        schema:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryImageAnalysisRequest"
        required: true
      responses:
        "200":
          description: Image marked for retry successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageAnalysisDetailsResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageAnalysisDetailsResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageAnalysisDetailsResponse"
  /api/admin/assessments/retry-all-failed:
    post:
      tags:
      - Admin Assessment Management
      summary: Bulk retry failed images
      description: "Marks all failed images across non-archived assessments as retry-pending,\
        \ limited by max retry count."
      operationId: bulkRetryFailedImages
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkRetryRequest"
        required: true
      responses:
        "200":
          description: Bulk retry operation completed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BulkOperationResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BulkOperationResponse"
  /api/admin/assessments/bulk-delete-by-status:
    post:
      tags:
      - Admin Assessment Management
      summary: Bulk archive by status
      description: Archives all non-archived assessments currently in the requested
        status.
      operationId: bulkDeleteByStatus
      parameters:
      - name: status
        in: query
        description: Status used to select assessments to archive
        required: true
        schema:
          type: string
          enum:
          - CREATED
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          - ARCHIVED
      responses:
        "200":
          description: Bulk archive operation completed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BulkOperationResponse"
        "400":
          description: Bad request (invalid status parameter)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BulkOperationResponse"
  /api/assessments/{assessmentId}:
    get:
      tags:
      - Assessments
      summary: Get assessment by ID
      description: "Returns a full assessment view including vehicle details, all\
        \ analyzed images, and generated report/quote paths."
      operationId: getAssessmentById
      parameters:
      - name: assessmentId
        in: path
        description: ID of the assessment
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Assessment returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetailsResponse"
        "404":
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetailsResponse"
...
  /api/admin/assessments:
    get:
      tags:
      - Admin Assessment Management
      summary: List assessments
      description: "Returns paginated admin assessment summaries. When status is omitted,\
        \ all active (non-archived) assessments are returned."
      operationId: getAssessmentsByStatus
      parameters:
      - name: status
        in: query
        description: Optional status filter
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          - ARCHIVED
      - name: page
        in: query
        description: Page number (0-indexed)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: Assessment page returned successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageAssessmentSummaryResponse"
        "400":
          description: Bad request (invalid paging or filter parameters)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageAssessmentSummaryResponse"
  /api/admin/assessments/{assessmentId}/images/{angle}:
    get:
      tags:
      - Admin Assessment Management
      summary: Get image analysis details
      description: "Returns analysis state for a specific image angle, including status,\
        \ token usage, retry count, and error information."
      operationId: getImageAnalysisDetails
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      - name: angle
        in: path
        description: Vehicle image angle to inspect
        required: true
        schema:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
      responses:
        "200":
          description: Image analysis details returned successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageAnalysisDetailsResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageAnalysisDetailsResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageAnalysisDetailsResponse"
...
components:
  schemas:
    UpdateAssessmentRequest:
      type: object
      properties:
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        odometerReading:
          type: string
        notes:
          type: string
      required:
      - vehicleDetails
    VehicleDetails:
      type: object
      properties:
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        modelYear:
          type: string
        colour:
          type: string
        engineNumber:
          type: string
        vinNumber:
          type: string
        vehicleCategory:
          type: string
        expiryDate:
          type: string
        extractedFields:
          type: object
          additionalProperties:
            type: string
    AssessmentSummaryResponse:
      type: object
      properties:
        assessmentId:
          type: string
        status:
          type: string
          enum:
          - CREATED
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          - ARCHIVED
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        totalImages:
          type: integer
          format: int32
        successfulImages:
          type: integer
          format: int32
        failedImages:
          type: integer
          format: int32
        totalTokensUsed:
          type: integer
          format: int64
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    VehicleDamageFinding:
      type: object
      properties:
        category:
          type: string
        severity:
          type: string
        summary:
          type: string
        panel:
          type: string
        recommendedAction:
          type: string
    RetryImageAnalysisRequest:
      type: object
      properties:
        angle:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
      required:
      - angle
    ImageAnalysisDetailsResponse:
      type: object
      properties:
        angle:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - SUCCESS
          - FAILED
          - RETRY_PENDING
        vehicleImagePath:
          type: string
        vehicleImageContentType:
          type: string
        tokensUsed:
          type: integer
          format: int64
        retryCount:
          type: integer
          format: int32
        errorMessage:
          type: string
        uploadedAt:
          type: string
          format: date-time
    BulkRetryRequest:
      type: object
      properties:
        maxRetries:
          type: integer
          format: int32
    BulkOperationResponse:
      type: object
      properties:
        totalProcessed:
          type: integer
          format: int32
        successCount:
          type: integer
          format: int32
        failureCount:
          type: integer
          format: int32
        message:
          type: string
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    AssessmentDetailsResponse:
      type: object
      properties:
        assessmentId:
          type: string
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        odometerReading:
          type: string
        totalTokensUsed:
          type: integer
          format: int64
        licenseDiscImagePath:
          type: string
        licenseDiscImageContentType:
          type: string
        reportHtmlPath:
          type: string
        reportPdfPath:
          type: string
        quoteHtmlPath:
          type: string
        quotePdfPath:
          type: string
        vehicleImages:
          type: array
          items:
            $ref: "#/components/schemas/VehicleImageResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VehicleImageResponse:
      type: object
      properties:
        angle:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
        redactionApplied:
          type: boolean
        tokensUsed:
          type: integer
          format: int64
        vehicleImagePath:
          type: string
        vehicleImageContentType:
          type: string
        findings:
          type: array
          items:
            $ref: "#/components/schemas/VehicleDamageFinding"
        uploadedAt:
          type: string
          format: date-time
    PageAssessmentSummaryResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentSummaryResponse"
        number:
          type: integer
          format: int32
        empty:
          type: boolean
...